Redesign: display only public users followings in following profile tab (#11051)
* Display only public users followings in following profile tab

* Force use of redesigned profile cell

* Fix followings counters to consider only users and groups and fix officialization_text

* Use link for personal url in profile

* Remove deprecated examples

* Change expected texts

* Adapt tests of follows

* Add REDESIGN_PENDING skip

* Add cell spec for redesigned_profile_cell

* Fix spelling

* Adapt tests to redesigned profile

* Remove deprecated examples

* Prepare profile actions buttons to use options

* Add missing resend email confirmation instructions button

* Display user_profile_bottom hook in redesigned profile cell

* Make profile_holder redesigned profile cell public available from hook

* Change styles of member_of user_profile partial

* Recover test

* simplify markup

* add separator

* Fix typo


---------

+++ decidim-core/app/cells/decidim/profile_actions_cell.rb
@@ -10,6 +10,7 @@ module Decidim
     ACTIONS_ITEMS = {
       edit_profile: { icon: "pencil-line", path: :profile_edit_path },
       create_user_group: { icon: "team-line", path: :profile_new_group_path },
+      resend_email_confirmation_instructions: { icon: "share-forward-line", path: :group_email_confirmation_path, options: { method: :post } },
       edit_user_group: { icon: "team-line", path: :edit_group_path },
       message: { icon: "mail-send-line", path: :new_conversation_path },
       disabled_message: { icon: "mail-send-line", options: { html_options: { disabled: true, title: I18n.t("decidim.user_contact_disabled") } } },
@@ -17,7 +18,7 @@ module Decidim
       manage_user_group_admins: { icon: "user-star-line", path: :profile_group_admins_path },
       invite_user: { icon: "user-add-line", path: :group_invites_path },
       join_user_group: { icon: "user-add-line", path: :group_join_requests_path, options: { method: :post } },
-      leave_user_group: { icon: "logout-box-r-line", path: :leave_group_path, options: { method: :delete } }
+      leave_user_group: { icon: "logout-box-r-line", path: :leave_group_path, options: { method: :delete, data: { confirm: I18n.t("decidim.groups.actions.are_you_sure") } } }
     }.freeze
 
     private
@@ -76,6 +77,7 @@ module Decidim
                                          :manage_user_group_admins,
                                          :invite_user
                                        ].tap do |keys|
+                                         keys.prepend(:resend_email_confirmation_instructions) if user_group_email_to_be_confirmed?
                                          keys << :join_user_group if can_join_user_group?
                                          keys << :leave_user_group if can_leave_group?
                                        end
@@ -122,6 +124,13 @@ module Decidim
       return false unless current_user
 
       !Decidim::UserGroupMembership.exists?(user: current_user, user_group: model)
+    end
+
+    def user_group_email_to_be_confirmed?
+      return false unless user_group?
+      return false unless current_user
+
+      !model.confirmed?
     end
 
     def group_member?
